Move integration tests into their own file
